refactor: deploy page with 3 steps
