+++ src/components/Deploy/1-DeployModeStep/index.tsx
@@ -1,0 +1,3 @@
+export function DeployModeStep() {
+  return <p>Deploy mode step</p>
+}
