+++ src/components/Deploy/2-ConfigurationStep/OperationsField/Operation/DisplayOperation.tsx
