+++ src/components/Deploy/2-ConfigurationStep/RestartFlied/index.tsx
@@ -1,5 +1,5 @@
 import { useDeployContext } from 'src/contexts'
-import { FieldHeader } from '../FieldHeader'
+import { FieldHeader } from '../../FieldHeader'
 import { DeployActionEnum } from 'src/types/deployTypes'
 
 export function RestartField() {
