+++ src/components/Deploy/3-ReviewStep/index.tsx
@@ -1,0 +1,3 @@
+export function ReviewStep() {
+  return <p>Review step</p>
+}
