+++ src/components/Deploy/index.tsx
@@ -1,5 +1,5 @@
-export * from './DeployTypeField'
+export * from './1-DeployModeStep'
+export * from './2-ConfigurationStep'
+export * from './3-ReviewStep'
 export * from './FieldHeader'
-export * from './FilterField'
-export * from './OperationsField'
-export * from './RestartFlied'
+export * from './ProgressBar'
